==> 0. Generating Splits.ipynb <==
import os
import json
import matplotlib
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm
from sklearn.model_selection import StratifiedGroupKFold, GroupKFold

DATA_PATH = "../raw_data/" # raw_data路径

topics = pd.read_csv(DATA_PATH + "topics.csv")
content = pd.read_csv(DATA_PATH + "content.csv")
correlations = pd.read_csv(DATA_PATH + "correlations.csv")

topics = topics[topics.has_content==True] # 只保留有content的topic

## Split

# 将所有category=="source"的topic，都划分到train fold中，因为隐藏测试集中没有category=="source"的topic
topics_train = topics[topics['category'] == "source"][["id"]] 
topics_train["fold"] = "train"

# valid df
topics_val = topics[topics['category'] != "source"].reset_index(drop=True)

# StratifiedGroupKFold
sgkf = StratifiedGroupKFold(random_state=1773,
                            n_splits=4,
                            shuffle=True)

# 根据language进行分层抽样，根据id进行分组
split_idxs = list(sgkf.split(topics_val["id"], topics_val["language"], groups=topics_val["id"]))[0]

split_idxs

# 创建训练集
topics_add_train = topics_val.iloc[split_idxs[0]].reset_index(drop=True)[["id"]]
topics_add_train["fold"] = "train"

topics_train = pd.concat([topics_train, topics_add_train], ignore_index=True)

# 创建验证集
topics_holdout = topics_val.iloc[split_idxs[1]].reset_index(drop=True)[["id"]]
topics_holdout["fold"] = "test"

# 查看训练集的language分布
topics[topics.id.isin(topics_train.id)].language.value_counts()

# 
topics[topics.id.isin(topics_holdout.id)].language.value_counts()

# 保存完整数据集
split_df = pd.concat([topics_train, topics_holdout], ignore_index=True)
split_df.to_csv('train_test_splits.csv', index=False)

Done !
