==> 2. Unsupervised Sampling.ipynb <==
import os
os.environ["KMP_DUPLICATE_LIB_OK"]="TRUE" # jupyter 环境设置

# =========================================================================================
# Libraries
# =========================================================================================
import gc
import time
import math
import random
import warnings
warnings.filterwarnings("ignore")
import numpy as np
import pandas as pd
from tqdm.auto import tqdm

# Custom libraries
from utils.unsupervised_utils import read_data, get_neighbors, build_training_set
from utils.utils import read_config
from utils.metrics import get_pos_score, get_f2_score

%env TOKENIZERS_PARALLELISM=false # tokens并行
%env TRANSFORMERS_NO_ADVISORY_WARNINGS=true # 关闭性能警告

config = read_config() # 读取配置文件
DATA_PATH = "../raw_data/"

# 对于 "test" 和 "train" 两种模式进行循环操作
for mode in [
    "test",
    "train"
]:
   
    print(f"FOR MODE - {mode}")
    
    # 读取数据
    topics, content, correlations, _ = read_data(data_path=DATA_PATH, config_obj=config, read_mode=mode)
    
    # 最近邻算法
    topics, content = get_neighbors(topic_df=topics, content_df=content, config_obj=config)
    
    # # 将 target 与 topic 进行合并，并计算 max positive score
    topics = topics.merge(correlations, how = 'inner', left_on = ['id'], right_on = ['topic_id'])
    pos_score = get_pos_score(topics['content_ids'], topics['predictions'])
    f2_score = get_f2_score(topics['content_ids'], topics['predictions'])
    print(f'Max Recall @{config["unsupervised_model"]["top_n"]}: {pos_score}')
    print(f'F2-Score @{config["unsupervised_model"]["top_n"]}: {f2_score}')
    
    # 设置 content 的索引为 id
    content.set_index('id', inplace = True)
    
    # 构建监督训练的训练集，在下一步中使用
    df = build_training_set(topic_df=topics, content_df=content)
    print(f'Our {mode} set has {len(df)} rows')
    
    df.to_parquet(f'./generated_files/unsupervised_{mode}.parquet')
    df.head()
    
    del df, topics, content, correlations
    gc.collect()
