==> 4. Inference.ipynb <==
# Sentence Transformers + Cross Encoders
## 2-Stage Inference Pipeline (Long Trained)

import os
os.environ["KMP_DUPLICATE_LIB_OK"]="TRUE"  # jupyter 环境设置

# 安装 sentence-transformers
!cp "/kaggle/input/sentence-transformers-lib-for-2023-01-23/sentence-transformers-2.2.2.piplib" "sentence-transformers-2.2.2.tar.gz" # 
!pip install sentence-transformers-2.2.2.tar.gz --no-dependencies -qqq

# =========================================================================================
# Libraries
# =========================================================================================
import gc
import time
import math
import random
import warnings
warnings.filterwarnings("ignore")
import numpy as np
import pandas as pd
from tqdm.auto import tqdm
import json

from cuml.neighbors import NearestNeighbors
import cupy as cp

from sentence_transformers import models, losses
from sentence_transformers import SentenceTransformer, CrossEncoder, util
from sentence_transformers.readers import InputExample
from sentence_transformers.cross_encoder.evaluation import CEBinaryClassificationEvaluator
from torch.utils.data import DataLoader
import torch

%env TOKENIZERS_PARALLELISM=true # tokens并行
%env TRANSFORMERS_NO_ADVISORY_WARNINGS=true # 关闭性能警告

def get_neighbors(topic_df,
                  content_df,
                  config_obj):
    
    # 创建无监督模型以提取 embeddings
    model = SentenceTransformer(config_obj["unsupervised_model"]["save_name"])
    model = model.to("cuda")

    # topics embeddings
    topics_preds = model.encode(topic_df["model_input"], show_progress_bar=True, convert_to_tensor=True)
    topics_preds_gpu = cp.asarray(topics_preds)

    # content embeddings
    content_preds = model.encode(content_df["model_input"], show_progress_bar=True, convert_to_tensor=True, batch_size=100)
    content_preds_gpu = cp.asarray(content_preds)

    # Release memory
    torch.cuda.empty_cache()
    gc.collect()

    # KNN model
    print(' ')
    print('Training KNN model...')
    # 定义KNN模型
    neighbors_model = NearestNeighbors(n_neighbors=config_obj["unsupervised_model"]["top_n"], metric='cosine')
    # 训练KNN模型
    neighbors_model.fit(content_preds_gpu)
    # 推理：获取最近的top_n个邻居
    indices = neighbors_model.kneighbors(topics_preds_gpu, return_distance=False)

    # 组成预测结果
    predictions = []
    for k in tqdm(range(len(indices))):
        pred = indices[k]
        p = ' '.join([content_df.loc[ind, 'id'] for ind in pred.get()])
        predictions.append(p)
    topic_df['predictions'] = predictions

    # Release memory
    del topics_preds, content_preds, topics_preds_gpu, content_preds_gpu, neighbors_model, predictions, indices, model
    gc.collect()
    torch.cuda.empty_cache()
    gc.collect()

    return topic_df, content_df



def build_training_set(topic_df,
                       content_df,
                       mode="local"):
    '''构建监督训练的训练数据集'''

    # 创建用于训练的列表
    topics_ids = []
    content_ids = []
    title1 = []
    title2 = []
    targets = []

    # 遍历每个 topic
    for k in tqdm(range(len(topic_df))):
        row = topic_df.iloc[k]
        topics_id = row['id']
        topics_title = row['model_input']
        predictions = row['predictions'].split(' ')
        
        # 如果模式为本地，则获取 ground truth
        if mode == "local":
            ground_truth = row['content_ids'].split(' ')

        for pred in predictions:
            content_title = content_df.loc[pred, 'model_input']
            topics_ids.append(topics_id)
            content_ids.append(pred)
            title1.append(topics_title)
            title2.append(content_title)

            # 如果pred在ground truth中，target为1，否则为0
            if mode == "local":
                if pred in ground_truth:
                    targets.append(1)
                else:
                    targets.append(0)
    
    # Build training DataFrame
    train = pd.DataFrame(
        {'topics_ids': topics_ids,
         'content_ids': content_ids,
         'model_input1': title1,
         'model_input2': title2
         }
    )
    if mode == "local":
        train["target"] = targets

    return train

def read_data(data_path,
              config_obj,
              read_mode="all"):
    topics = pd.read_csv(data_path + 'topics.csv')
    content = pd.read_csv(data_path + 'content.csv')

    if read_mode != "all":
        correlations = pd.read_csv(data_path + 'correlations.csv')
    else:
        correlations = None
    topic_trees = generate_topic_tree(topics) # 生成主题树df

    if read_mode != "all":
        splits = pd.read_csv("train_test_splits.csv")  # 读取训练和测试数据的split情况
        topics = topics[topics.id.isin(splits[splits.fold == read_mode].id)].reset_index(drop=True) # 根据read_mode筛选train/valid的主题

    topics = topics.merge(topic_trees, how="left", on="id")  # 将主题数据与主题树合并
    del topic_trees # 删除主题树变量以释放内存
    gc.collect()

    topic_tokens = generate_topic_model_input(input_df=topics) # 生成主题模型的输入数据
    content_tokens = generate_content_model_input(input_df=content) # 生成内容模型的输入数据

    unq_tokens = set(topic_tokens + content_tokens + ["nan"]) # 所有唯一的tokens

    # 按 title 长度排序，以加快推理过程
    topics['length'] = topics['title'].apply(lambda x: len(x)) # 主题 title 长度
    content['length'] = content['title'].apply(lambda x: len(x)) # 内容 title 长度
    topics.sort_values('length', inplace=True) # 对主题按照 title长度进行排序
    content.sort_values('length', inplace=True) # 对内容按照 title长度进行排序

    # 删除列
    topics.drop(['length'], axis=1, inplace=True)
    content.drop(['length'], axis=1, inplace=True)

    # 重置索引
    topics.reset_index(drop=True, inplace=True)
    content.reset_index(drop=True, inplace=True)
    print(' ')
    print('-' * 50)
    print(f"topics.shape: {topics.shape}")
    print(f"content.shape: {content.shape}")
    if read_mode != "all":
        print(f"correlations.shape: {correlations.shape}")

    return topics, content, correlations, unq_tokens


def generate_token_features(input_df, token_features):
    """
    Args:
        input_df
        token_features: 需要生成token的列名list
    Returns:
        Tuple of (带有额外模型输入列的df, Unique Special Tokens)
        Tuple of (Dataframe with additional model input column, Unique Special Tokens)
    """
    token_feature_set = None # 初始化token特征集
    special_tokens = [] # 初始化 sep token 列表
    # 遍历 token_features 中的每个特征列
    for token_feature in token_features:
        # 为当前特征生成 sep token 字符串
        token_feature_str = "[<[" + token_feature + "_" + input_df[token_feature].astype(str) + "]>]"
        special_tokens += set(token_feature_str.values)  # 将生成的 sep token 添加到列表中
        
        # 将生成的 sep token 字符串添加到 token_feature_set 中
        if not isinstance(token_feature_set, pd.Series):
            token_feature_set = token_feature_str
        else:
            token_feature_set += " " + token_feature_str

    # 返回 token_feature_set 和 Unique Special Tokens
    return token_feature_set, special_tokens


def generate_topic_model_input(input_df):
    """
    Args:
        input_df
    Returns:
        修改 input_df
        Dataframe with additional model input column
    """
    input_df.fillna("nan", inplace=True) # 用"nan"填充空值

    # token特征列表
    token_features = ["language",]

    # 生成 token_feature_text 和 Special Tokens
    token_feature_text, special_tokens = generate_token_features(input_df=input_df, token_features=token_features)

    # 生成 model_input，将 token_feature_text、title、topic_tree 和 description 组合起来，并将所有文本转换为小写
    input_df["model_input"] = (
            token_feature_text +
            " [<[topic_title]>] " + input_df["title"].astype(str) +
            " [<[topic_tree]>] " + input_df["topic_tree"].astype(str) +
            " [<[topic_desc]>] " + input_df["description"].astype(str)
    ).str.lower()#.str.split().apply(lambda x: " ".join(x[:seq_len]))

    del token_feature_text

    input_df.drop(['description', 'channel', 'category',
                   'level', 'parent', 'has_content'],
                  axis=1,
                  inplace=True)
    gc.collect()

    # 返回 special_tokens, input_df 在函数内部被修改
    return special_tokens


def generate_content_model_input(input_df):
    """
    Args:
        input_df
    Returns:
        修改 input_df
        Dataframe with additional model input column
    """

    input_df.fillna("nan", inplace=True) # 用"nan"填充空值

    # token特征列表
    token_features = ["language", "kind"]

    # 生成 token_feature_text 和 Special Tokens
    token_feature_text, special_tokens = generate_token_features(input_df=input_df, token_features=token_features)

    # 生成 model_input，将 token_feature_text、title、description 和 text 组合起来，限制长度为512个单词，并将所有文本转换为小写
    input_df["model_input"] = (
            token_feature_text +
            " [<[cntnt_title]>] " + input_df["title"].astype(str) +
            " [<[cntnt_desc]>] " + input_df["description"].astype(str) +
            " [<[cntnt_text]>] " + input_df["text"].astype(str)
    ).apply(lambda x: " ".join(x.split()[:512])).str.lower()

    del token_feature_text

    input_df.drop(['description', 'kind', 'text', 'copyright_holder', 'license'],
                  axis=1,
                  inplace=True)
    gc.collect()

    # 返回 special_tokens, input_df 在函数内部被修改
    return special_tokens


def generate_topic_tree(input_topic_df):
    '''
    生成主题树
    Args:
        input_topic_df: topic_df
    Returns:
        df: 包含主题树的 DataFrame, 其中包含列: "id", "topic_tree", "reverse_level"
    '''
    df = pd.DataFrame()

    # 遍历所有唯一的channel
    for channel in tqdm(input_topic_df['channel'].unique()):
        # 筛选出当前 channel 的数据，并重置索引
        channel_df = input_topic_df[(input_topic_df['channel'] == channel)].reset_index(drop=True)
        # 对当前 channel 的 level 进行排序
        for level in sorted(channel_df.level.unique()):
            # 对于 level 为 0 的情况，先为该主题创建一个 topic tree 列，该列为该主题的 title
            if level == 0:
                topic_tree = channel_df[channel_df['level'] == level]['title'].astype(str)
                topic_tree_df = pd.DataFrame([channel_df[channel_df['level'] == level][['id']], topic_tree.values]).T
                # id -> child_id, topic_tree -> topic_tree
                topic_tree_df.columns = ['child_id', 'topic_tree']
                # 将主题树与原始数据合并
                channel_df = channel_df.merge(topic_tree_df, left_on='id', right_on='child_id', how='left').drop(['child_id'], axis=1)

            # 创建 topic tree 列后，将父节点和子节点合并，其中 parent_id = child_id
            topic_df_parent = channel_df[channel_df['level'] == level][['id', 'title', 'parent', 'topic_tree']]
            topic_df_parent.columns = 'parent_' + topic_df_parent.columns

            topic_df_child = channel_df[channel_df['level'] == level + 1][['id', 'title', 'parent', 'topic_tree']]
            topic_df_child.columns = 'child_' + topic_df_child.columns

            # 合并父子节点，生成 topic_df_merged
            topic_df_merged = topic_df_parent.merge(topic_df_child, left_on='parent_id', right_on='child_parent')[
                ['child_id', 'parent_id', 'parent_title', 'child_title', 'parent_topic_tree']]

            # topic tree 是父 topic tree 加上当前子级的 title: 'parent > child'
            topic_tree = topic_df_merged['parent_topic_tree'].astype(str) + ' > ' + topic_df_merged[
                'child_title'].astype(str)

            # 创建包含子节点 id 和 topic_tree 的 DataFrame
            topic_tree_df = pd.DataFrame([topic_df_merged['child_id'].values, topic_tree.values]).T
            topic_tree_df.columns = ['child_id', 'topic_tree']

            # 将 topic_tree_df 与 channel_df 合并，并删除 'child_id' 列
            channel_df = channel_df.merge(topic_tree_df, left_on='id', right_on='child_id', how='left').drop(['child_id'], axis=1)

            # 如果存在 'topic_tree_y' 列，则将其与 'topic_tree_x' 列合并
            if 'topic_tree_y' in list(channel_df.columns):
                # combine_first: 用非空值填充空值, 如果两个Series 都不为空,优先使用前者。
                channel_df['topic_tree'] = channel_df['topic_tree_x'].combine_first(channel_df['topic_tree_y'])
                channel_df = channel_df.drop(['topic_tree_x', 'topic_tree_y'], axis=1)
        
        # 将当前 channel 的数据添加到总的 DataFrame 中
        df = pd.concat([df, channel_df], ignore_index=True)

    # 计算每个 channel 的最大 level，并添加到 DataFrame
    df = df.merge(df.groupby("channel")["level"].max().rename("max_level_of_channel").reset_index(),
                  how="left",
                  on="channel")
    
    # 逆序 level
    df["reverse_level"] = df["max_level_of_channel"] - df["level"]

    return df[["id", "topic_tree", "reverse_level"]]

config = {
  "unsupervised_model": {
    "save_name": "trained_models/unsupervised/paraphrase-multilingual-mpnet-base-v2",
    "seq_len": 128,
    "top_n": 100, # knn top n
    "seed": 42
  },
  "supervised_model": {
    "save_name": "trained_models/supervised/paraphrase-multilingual-mpnet-base-v2",
    "threshold": 0.015, # threshold for positive sample
    "seq_len": 128
  }
}

DATA_PATH = "../raw_data/"

submission_df = pd.read_csv(DATA_PATH + "sample_submission.csv")
submission_df.head()

## 1. Retrieve

# 读取全量数据
topics, content, correlations, _  = read_data(data_path=DATA_PATH, config_obj=config, read_mode="all") 
# 只保留出现在 submission_df 中的topic
topics = topics[topics.id.isin(submission_df.topic_id)].reset_index(drop=True)

# 最近邻算法
topics, content = get_neighbors(topic_df=topics, content_df=content, config_obj=config)

# 为 content 设置 id 为索引
content.set_index('id', inplace = True)

# 构建监督训练的训练集，在下一步中使用
df = build_training_set(topic_df=topics, content_df=content, mode="kaggle")

df.head()

del topics, content

## 2. Re-Rank

# 创建一个 CrossEncoder 模型，并加载训练后的模型
model = CrossEncoder(config["supervised_model"]["save_name"], # 模型名称
                    num_labels=1,
                    max_length=config["supervised_model"]["seq_len"])

preds = model.predict(df[["model_input1", "model_input2"]].values, # 测试数据
                      show_progress_bar=True,
                      batch_size=96
                      )

df["pred_score"] = preds # 保存预测结果

df.head()

sorted_pred_df = df.sort_values(by="pred_score", ascending=False).reset_index(drop=True) # 按照预测分数降序排序

# 根据 thr 筛选出预测值大于等于thr的数据
# 按照 topics_ids 对筛选后的数据进行分组，并将每个组内的 content_ids 进行拼接，然后重命名结果列为 "pred_content_ids"。
preds_thr_df = sorted_pred_df[["topics_ids","content_ids"]][sorted_pred_df.pred_score >= config["supervised_model"]["threshold"]].\
                                groupby("topics_ids")["content_ids"].apply(lambda x: " ".join(x)).rename("pred_content_ids").reset_index()

## Create Submission

# 将结果合并到 submission_df 中
submission_df = submission_df.merge(preds_thr_df,
                                    how="left",
                                    right_on="topics_ids",
                                    left_on="topic_id")[["topic_id", "pred_content_ids"]]
submission_df.rename(columns={"pred_content_ids": "content_ids"}, inplace=True)

submission_df

submission_df.to_csv("submission.csv", index=False)
